+++ app/containers/Status/definition.ts
@@ -1,6 +1,6 @@
 import { TextProps } from 'react-native';
 
-import { TUserStatus } from '../../definitions/UserStatus';
+import { TUserStatus } from '../../definitions/TUserStatus';
 
 export interface IStatus extends TextProps {
 	id: string;
