+++ app/containers/Status/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 
-import { TUserStatus } from '../../definitions/UserStatus';
+import { TUserStatus } from '../../definitions/TUserStatus';
 import { IApplicationState } from '../../definitions';
 import Status from './Status';
 import { IStatus } from './definition';
