+++ app/lib/rocketchat/services/connect.ts
@@ -6,7 +6,7 @@ import { Q } from '@nozbe/watermelondb';
 
 import log from '../../../utils/log';
 import { onRolesChanged } from '../../methods/getRoles';
-import { UserStatus } from '../../../definitions/UserStatus';
+import { STATUSES } from '../../../definitions/TUserStatus';
 import { setActiveUsers } from '../../../actions/activeUsers';
 import protectedFunction from '../../methods/helpers/protectedFunction';
 import database from '../../database';
@@ -17,7 +17,7 @@ import { store } from '../../auxStore';
 import { loginRequest, setLoginServices, setUser } from '../../../actions/login';
 import sdk from './sdk';
 import I18n from '../../../i18n';
-import RocketChat, { MIN_ROCKETCHAT_VERSION, STATUSES } from '../rocketchat';
+import RocketChat, { MIN_ROCKETCHAT_VERSION } from '../rocketchat';
 import { ICredentials, ILoggedUser, IRocketChat } from '../../../definitions';
 import { isIOS } from '../../../utils/deviceInfo';
 import { connectRequest, connectSuccess, disconnect as disconnectAction } from '../../../actions/connect';
@@ -195,7 +195,7 @@ function connect(
 
 					const { user: loggedUser } = store.getState().login;
 					if (loggedUser && loggedUser.id === id) {
-						store.dispatch(setUser({ status: STATUSES[status] as UserStatus, statusText }));
+						store.dispatch(setUser({ status: STATUSES[status], statusText }));
 					}
 				} else if (/updateAvatar/.test(eventName)) {
 					const { username, etag } = ddpMessage.fields.args[0];
