+++ app/reducers/activeUsers.test.ts
@@ -1,5 +1,4 @@
 import { clearActiveUsers, setActiveUsers } from '../actions/activeUsers';
-import { UserStatus } from '../definitions/UserStatus';
 import { IActiveUsers, initialState } from './activeUsers';
 import { mockedStore } from './mockedStore';
 
@@ -9,7 +8,7 @@ describe('test reducer', () => {
 		expect(state).toEqual(initialState);
 	});
 	it('should return modified store after action', () => {
-		const activeUsers: IActiveUsers = { any: { status: UserStatus.ONLINE, statusText: 'any' } };
+		const activeUsers: IActiveUsers = { any: { status: 'online', statusText: 'any' } };
 		mockedStore.dispatch(setActiveUsers(activeUsers));
 		const state = mockedStore.getState().activeUsers;
 		expect(state).toEqual({ ...activeUsers });
